+++ backend/MarketApi/MarketApi/Controllers/ProductController.cs
@@ -6,11 +6,10 @@ using System.Security.Claims;
 
 namespace MarketApi.Controllers
 {
+    [ApiController]
+    [Route("api/[controller]")]
     public class ProductController : Controller
     {
-        public IActionResult Index()
-        {
-            return View();
-        }
+        
     }
 }