+++ backend/MarketApi/MarketApi/Migrations/20250805153633_AllowMultipleProductFiles.cs
@@ -1,0 +1,64 @@
+﻿using System;
+using System.Collections.Generic;
+using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace MarketApi.Migrations
+{
+    /// <inheritdoc />
+    public partial class AllowMultipleProductFiles : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<DateTime>(
+                name: "CreatedAt",
+                table: "Products",
+                type: "timestamp with time zone",
+                nullable: false,
+                defaultValue: new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified));
+
+            migrationBuilder.AddColumn<string>(
+                name: "Currency",
+                table: "Products",
+                type: "character varying(10)",
+                maxLength: 10,
+                nullable: false,
+                defaultValue: "");
+
+            migrationBuilder.AddColumn<List<string>>(
+                name: "FilePaths",
+                table: "Products",
+                type: "text[]",
+                nullable: false);
+
+            migrationBuilder.AddColumn<decimal>(
+                name: "Price",
+                table: "Products",
+                type: "numeric",
+                nullable: false,
+                defaultValue: 0m);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "CreatedAt",
+                table: "Products");
+
+            migrationBuilder.DropColumn(
+                name: "Currency",
+                table: "Products");
+
+            migrationBuilder.DropColumn(
+                name: "FilePaths",
+                table: "Products");
+
+            migrationBuilder.DropColumn(
+                name: "Price",
+                table: "Products");
+        }
+    }
+}