+++ backend/MarketApi/MarketApi/Models/AppDbContext.cs
@@ -7,6 +7,7 @@ namespace MarketApi.Models
     {
         public DbSet<User> Users { get; set; }
         public DbSet<Product> Products { get; set; }
+        public DbSet<ProductFile> ProductFiles { get; set; }
 
 
         public AppDbContext(DbContextOptions<AppDbContext> options) : base(options) { }