+++ backend/MarketApi/MarketApi/Services/IProductService.cs
@@ -1,0 +1,10 @@
+﻿using MarketApi.Models;
+
+namespace MarketApi.Services
+{
+    public interface IProductService
+    {
+        Product GetProductResponse(Product product);
+        ProductFile GetProductFileResponse(ProductFile productFile);
+    }
+}